Flyttet logging av responstid et nivå opp
Endret på hvor vi starter og stopper tiden etter samtale med @kjetillvassmolund.

Eks på loggmelding: ResponseTime: DIALOG_INNBYGGER_EKONTAKT: FromHerId: 93252 ToHerId: 93238 MessageId: 03734f70-52ae-4840-8b7c-066849bc50f3 ResponseTime: 53 ms
